// src/app/interfaces/plantilla-concar.ts
export interface PlantillaCONCAR {
  Sub_Diario: string;
  Número_de_Comprobante: string;
  Fecha_de_Comprobante: string;
  Código_de_Moneda: string;
  Glosa_Principal: string;
  Tipo_de_Cambio: string;
  Tipo_de_Conversión: string;
  Flag_de_Conversión_de_Moneda: string;
  Fecha_Tipo_de_Cambio: string;
  Cuenta_Contable: string;
  Código_de_Anexo: string;
  Código_de_Centro_de_Costo: string;
  Debe_____Haber: string;
  Importe_Original: string;
  Importe_en_Dólares: string;
  Importe_en_Soles: string;
  Tipo_de_Documento: string;
  Número_de_Documento: string;
  Fecha_de_Documento: string;
  Fecha_de_Vencimiento: string;
  Código_de_Area: string;
  Glosa_Detalle: string;
  Código_de_Anexo_Auxiliar: string;
  Medio_de_Pago: string;
  Tipo_de_Documento_de_Referencia: string;
  Número_de_Documento_Referencia: string;
  Fecha_Documento_Referencia: string;
  Nro_Máq__Registradora_Tipo_Doc__Ref_: string;
  Base_Imponible_Documento_Referencia: string;
  IGV_Documento_Provisión: string;
  Tipo_Referencia_en_estado_MQ: string;
  Número_Serie_Caja_Registradora: string;
  Fecha_de_Operación: string;
  Tipo_de_Tasa: string;
  Tasa_Detracción_____Percepción: string;
  Importe_Base_Detracción_____Percepción_Dólares: string;
  Importe_Base_Detracción_____Percepción_Soles: string;
  Tipo_Cambio_para_____F___: string;
  Importe_de_IGV_sin_derecho_crédito_fiscal: string;
}